requirements.txt
aniso8601==9.0.1
asgiref==3.5.0
cffi==1.15.0
click==8.0.4
cryptography==36.0.2
cycler==0.11.0
Django==4.0.3
Flask==2.0.3
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
fonttools==4.31.2
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.1
Jinja2==3.0.3
kiwisolver==1.4.2
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
matplotlib==3.5.1
mysql-client==0.0.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.22.3
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
protobuf==3.19.4
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
schedule==1.1.0
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
Werkzeug==2.0.3
whitenoise==6.0.0
